uploaded file
a work in progress. Basically I updated Discord.js and it ruined my old
bot so now I have to account for all the changes and start from scratch.
(I was using a REALLY out of date version) Feel free to use this as a
reference for your own bot as Discord.js documentation for the latest
versions aren't as thorough as they used to be. Also, Node.js is
required, there are more details in the program comments. I'll make
installation instructions once I'm satisfied with this project.